redirect all stderr writes into /dev/null
This directs all writes to stderr into /dev/null.

This is only done because ffmpeg, through libffmpegthumbnailer,
will output its warnings/errors when files are having their thumbnails
generated. Which floods stderr with messages that the user can do nothing
about, such as:

'deprecated pixel format used, make sure you did set range correctly'

An alternative solution to this would be to use
Glib::spawn_command_line_sync and redirecting that output to
/dev/null, but that would involve using the libffmpegthumbnailer CLI
program and not the C++ interface. Not a solution that I want to do.