wayland fix DnD always choosing to copy
The problem was that when using wayland gdk_drag_status() was not
updating drag_context with the correct action. Worked fine under X11.

Before this was calling stat() on every drag event, now only calls
stat() when the DnD event is over and needs to determine if the action
is a copy or move.