remove plugin support
This removes all plugin support.

The plugin code was virtually unmaintained and a maintenance burden. The
only practical use for plugins was xdg trash support, but I added native
support a while ago. Any additional features that a plugin could be used
for can be written in cpp and put into the main repo.