rewrite archive create/extract/list
external archive managers are and will always be better than any
internal implementation, so just use one of them.

this uses file-roller as the archiver, xarchive was not choosen because
from the CLI it will only extract the first file in the list and ignore
the others, while file-roller will extract all the files in the list.