PtkFileBrowser update selection history logic
Selection history currently gets maintained separately from the
navigaion history. Once the navigation history is ported away from
using a GList then the selection history can then be slaved to the
navigation history.

This requires a hack/abstraction layer using selection_history_data to
work with the PtkFileBrowser GType. The problem is that without the
selection_history_data abstraction accessing the selection history
std::map will immediately sefgault. The GType is doing something bad
and I have no idea what it is so simply work around it.

There is a similar problem with the VFSDir cache global std::map where
the key has to be a char* or it will break multiple tabs using the
same VFSDir.