PtkFileBrowser chdir/refresh load vfs::dir cleanup
in chdir() need to disconnect form the signal before setting a new
signal. do not call refresh() in chdir(), this was causing loading
problems.

in refresh() do not set the dir_ variable again since it was set
with the correct vfs::dir object in chdir(). also do not connect
signals in refresh() they are already connected in chdir().