import natsort implementation
Another sorting change.

Problems with other sorting methods. Using std::collate or
Glib::ustring::collate_key() causes numbers to not be sorted in a human
friendly way i.e. sorts as (1, 10, 2, ...), not as (1, 2, 10, ...). Using
the Glib g_utf8_collate_* functions will sort number correctly but causes
symbols to be sorted inline with letters, i.e. sorts as ([a, a, [b, b),
not as (a, b, [a, [b).