vfs::dir::refresh() reload directory/files
commit c9aaf68ae19a2c88e913a4eb4da0ec2ce339a7c2 "broke" dir refreshing.

When vfs::dir used a GType, the dir cache was broken and would create a
new vfs::dir for each refresh, this was broken but had the same effect
as a refresh. Now that the dir cache is not broken refreshing no longer
worked, but is fixed now. Refreshing now rereads the dir and adds missing
files, removes/adds files added/removed to the .hidden file, reloads/creates
new thumbnails.