compile time option to build without exo
GTK only builds are runtime broken.

With the GTK build, mouse DnD will only try and operate on a single
file even when multiple are selected. Multile item selection works
correctly with the Copy/Paste/Trash/Delete keybindings and context
menus.

I have worked in this on and off and have not been able to fix it, so
add a compile time option so this can be easily worked on at a later
time.

exo is also gtk3 only and is in the way of the gtk4 port.