vfs::dir rewrite to use concurrencpp
This removes vfs::async_thread which was a cpp 'port' of vfs::async_task
that used std::thread. It was supposed to replace all usage of
vfs::async_task but only vfs::dir ended up using it. The main problems
with vfs::async_task was that it blocks. This could have been fixed by
using std::thread::detach(), but it would have still been a mess.