Initial Gtk4 rewrite
This is very much a work in progress and is not currently intended or
usable for general consumption, although it does work.

This is a rewrite of the gui from Gtk3 to Gtk4. The original plan was
to port the existing Gtk3 C code to Gtk4 C code and then do a piecemeal
port to Gtkmm4 C++. That proved to be a non viable option for multiple
reasons. Existing code, logic, and design are reused when possible but
there are behavior changes and this is not intend to be a one-to-one
feature port.